# docs/swagger.yaml
definitions:
...
  broadcast.ReadParam:
    properties:
      broadcast_id:
        type: string
    required:
    - broadcast_id
    type: object
  broadcast.SendBroadcastParam:
    properties:
      content:
        type: string
      scope:
        description: ALL | BUSINESS | CONSUMER
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
info:
  contact:
    name: Charlie
    url: https://github.com/jiangbyte
  description: Hei Gin 后台管理系统 API 文档
  license:
    name: MIT
    url: https://github.com/jiangbyte/hei-gin/blob/main/LICENSE
  termsOfService: https://github.com/jiangbyte/hei-gin
  title: Hei Gin API
  version: 1.0.0
paths:
...
  /api/v1/c/im/broadcast/detail:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/broadcast/detail，即时通讯广播详情查询
      parameters:
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯广播详情查询
      tags:
      - 即时通讯广播
  /api/v1/c/im/broadcast/read:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/c/im/broadcast/read，即时通讯广播标记已读
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/broadcast.ReadParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯广播标记已读
      tags:
      - 即时通讯广播
...
  /api/v1/sys/im/broadcast/detail:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/broadcast/detail，即时通讯广播详情查询
      parameters:
      - description: id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯广播详情查询
      tags:
      - 即时通讯广播
  /api/v1/sys/im/broadcast/list:
    get:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/broadcast/list，即时通讯广播列表查询
      parameters:
      - description: cursor
        in: query
        name: cursor
        type: string
      - description: size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯广播列表查询
      tags:
      - 即时通讯广播
  /api/v1/sys/im/broadcast/read:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/broadcast/read，即时通讯广播标记已读
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/broadcast.ReadParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯广播标记已读
      tags:
      - 即时通讯广播
  /api/v1/sys/im/broadcast/send:
    post:
      consumes:
      - application/json
      description: 访问 /api/v1/sys/im/broadcast/send，即时通讯广播发送消息
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/broadcast.SendBroadcastParam'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            additionalProperties: true
            type: object
      summary: 即时通讯广播发送消息
      tags:
      - 即时通讯广播
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"